# _apidocs/swagger.yaml
basePath: /
definitions:
  dto.BlogResponse:
    properties:
      body:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  dto.CreateBlogRequest:
    properties:
      body:
        type: string
      description:
        type: string
      title:
        type: string
    required:
    - body
    - title
    type: object
  dto.PaginatedBlogResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.BlogResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.UpdateBlogRequest:
    properties:
      body:
        type: string
      description:
        type: string
      title:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: This is a server for Go Blog API.
  title: Go Blog API
  version: "1.0"
paths:
  /api/blogs:
    get:
      consumes:
      - application/json
      description: Get all blogs with pagination
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedBlogResponse'
      summary: Get blogs
      tags:
      - blogs
    post:
      consumes:
      - application/json
      description: Create a new blog
      parameters:
      - description: Blog details
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBlogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlogResponse'
      summary: Create a blog
      tags:
      - blogs
  /api/blogs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a blog post by its ID
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a blog
      tags:
      - blogs
    get:
      consumes:
      - application/json
      description: Get a blog post by its ID
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlogResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a single blog
      tags:
      - blogs
    put:
      consumes:
      - application/json
      description: Update a blog post by its ID
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: integer
      - description: Blog update object
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateBlogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlogResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a blog
      tags:
      - blogs
swagger: "2.0"